download tast přesunut do vlastního souboru
